move files, deprecate isExperienceDismissed
